requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
attrs==21.4.0
Babel==2.9.1
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.5
django-colorfield==0.7.1
django-debug-toolbar==3.4.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pillow==9.1.1
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
yarl==1.7.2
